requirements.txt
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
streamlit==1.28.0
faiss-cpu==1.7.4
streamlit-extras==0.3.4
langchain==0.0.154
PyPDF2==3.0.1
python-dotenv==1.0.0
#altair==4.1.0
openai
tiktoken
pip install -r requirements.txt

